+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Changed
  - Faster Septentrio Mosaic-X5 detection
 ### Fixed
+ - Fix some services were not restarted after saving new settings in the main service.
  - Fix pystemd result request for timer services. #162 #445
  - Increase free space for archiving Mosaic-X5 data. #369
  - Various fixes #441 #443